feat(auth): replace next-auth middleware with optimistic Better Auth gate
Phase 1: convert middleware.js -> middleware.ts. The next-intl locale-routing composition is unchanged; next-auth's withAuth is replaced by an optimistic getSessionCookie() existence check (better-auth/cookies) that redirects to /login when no session cookie is present. Per Better Auth docs this is not a security boundary — real admin enforcement stays in the dashboard layout.

Part of #81
