--- middleware.js
@@ -1,49 +1,0 @@
-import { withAuth } from 'next-auth/middleware';
-import createIntlMiddleware from 'next-intl/middleware';
-import { NextRequest } from 'next/server';
-
-const locales = ['en', 'tr'];
-const publicPages = ['/', '/login', '/works', '/contact', '/about'];
-
-const intlMiddleware = createIntlMiddleware({
-  locales,
-  localePrefix: 'as-needed',
-  defaultLocale: 'en'
-});
-
-const authMiddleware = withAuth(
-  // Note that this callback is only invoked if
-  // the `authorized` callback has returned `true`
-  // and not for pages listed in `pages`.
-  function onSuccess(req) {
-    return intlMiddleware(req);
-  },
-  {
-    callbacks: {
-      authorized: ({ token }) => token != null
-    },
-    pages: {
-      signIn: '/login'
-    }
-  }
-);
-
-export default function middleware(req) {
-  const publicPathnameRegex = RegExp(
-    `^(/(${locales.join('|')}))?(${publicPages
-      .flatMap((p) => (p === '/' ? ['', '/'] : p))
-      .join('|')})/?$`,
-    'i'
-  );
-  const isPublicPage = publicPathnameRegex.test(req.nextUrl.pathname);
-
-  if (isPublicPage) {
-    return intlMiddleware(req);
-  } else {
-    return authMiddleware(req);
-  }
-}
-
-export const config = {
-  matcher: ['/((?!api|_next|.*\\..*).*)']
-};
